+++ src/snmp/lldpRemOrgDefInfoTable_interface.c
@@ -247,10 +247,10 @@ int lldpRemOrgDefInfoTable_index_from_oid(
 
     err = parse_oid_indexes(oid_idx->oids, oid_idx->len, &var_lldpRemTimeMark);
     if (err == SNMP_ERR_NOERROR) {
-        mib_idx->lldpRemTimeMark = *((long *)var_lldpRemTimeMark.val.string);
+        mib_idx->lldpRemTimeMark = *((long *)((uintptr_t)var_lldpRemTimeMark.val.string));
         mib_idx->lldpRemLocalPortNum =
-            *((long *)var_lldpRemLocalPortNum.val.string);
-        mib_idx->lldpRemIndex = *((long *)var_lldpRemIndex.val.string);
+            *((long *)((uintptr_t)var_lldpRemLocalPortNum.val.string));
+        mib_idx->lldpRemIndex = *((long *)((uintptr_t)var_lldpRemIndex.val.string));
         if (var_lldpRemOrgDefInfoOUI.val_len >
             sizeof(mib_idx->lldpRemOrgDefInfoOUI)) {
             err = SNMP_ERR_GENERR;
@@ -263,9 +263,9 @@ int lldpRemOrgDefInfoTable_index_from_oid(
                 sizeof(mib_idx->lldpRemOrgDefInfoOUI[0]);
         }
         mib_idx->lldpRemOrgDefInfoSubtype =
-            *((long *)var_lldpRemOrgDefInfoSubtype.val.string);
+            *((long *)((uintptr_t)var_lldpRemOrgDefInfoSubtype.val.string));
         mib_idx->lldpRemOrgDefInfoIndex =
-            *((long *)var_lldpRemOrgDefInfoIndex.val.string);
+            *((long *)((uintptr_t)var_lldpRemOrgDefInfoIndex.val.string));
     }
 
     snmp_reset_var_buffers(&var_lldpRemTimeMark);
